MAPSJS-2660: Move getBrowserLanguages to harp-utils.
