+++ @here/harp-utils/lib/DOMUtils.ts
@@ -1,0 +1,33 @@
+/*
+ * Copyright (C) 2021 HERE Europe B.V.
+ * Licensed under Apache 2.0, see full license in LICENSE
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export namespace DOMUtils {
+    /**
+     * Gets language list used by the browser
+     *
+     * @returns Array of iso language codes
+     */
+    export function getBrowserLanguages(): string[] | undefined {
+        if (navigator.languages !== undefined && navigator.languages.length > 0) {
+            const languageList = [];
+            for (const lang of navigator.languages) {
+                languageList.push(getIsoLanguageCode(lang));
+            }
+            return languageList;
+        }
+        if (navigator.language !== undefined) {
+            return [getIsoLanguageCode(navigator.language)];
+        }
+        return undefined;
+    }
+
+    /**
+     * Gets ISO-639-1 language code from browser's code (ex. en for en-US)
+     */
+    function getIsoLanguageCode(language: string) {
+        return language.substring(0, 2);
+    }
+}
